moved some things around to match common standards
